Testing split of monitor
